+++ src/Delimiter/Delimiter.php
@@ -16,37 +16,37 @@ namespace League\CommonMark\Delimiter;
 
 use League\CommonMark\Inline\Element\AbstractStringContainer;
 
-class Delimiter
+final class Delimiter implements DelimiterInterface
 {
     /** @var string */
-    protected $char;
+    private $char;
 
     /** @var int */
-    protected $numDelims;
+    private $length;
 
     /** @var int */
-    protected $origDelims;
+    private $originalLength;
 
     /** @var AbstractStringContainer */
-    protected $inlineNode;
+    private $inlineNode;
 
-    /** @var Delimiter|null */
-    protected $previous;
+    /** @var DelimiterInterface|null */
+    private $previous;
 
-    /** @var Delimiter|null */
-    protected $next;
-
-    /** @var bool */
-    protected $canOpen;
+    /** @var DelimiterInterface|null */
+    private $next;
 
     /** @var bool */
-    protected $canClose;
+    private $canOpen;
 
     /** @var bool */
-    protected $active;
+    private $canClose;
+
+    /** @var bool */
+    private $active;
 
     /** @var int|null */
-    protected $index;
+    private $index;
 
     /**
      * @param string                  $char
@@ -59,8 +59,8 @@ class Delimiter
     public function __construct(string $char, int $numDelims, AbstractStringContainer $node, bool $canOpen, bool $canClose, ?int $index = null)
     {
         $this->char = $char;
-        $this->numDelims = $numDelims;
-        $this->origDelims = $numDelims;
+        $this->length = $numDelims;
+        $this->originalLength = $numDelims;
         $this->inlineNode = $node;
         $this->canOpen = $canOpen;
         $this->canClose = $canClose;
@@ -69,7 +69,7 @@ class Delimiter
     }
 
     /**
-     * @return bool
+     * {@inheritdoc}
      */
     public function canClose(): bool
     {
@@ -77,19 +77,15 @@ class Delimiter
     }
 
     /**
-     * @param bool $canClose
-     *
-     * @return $this
+     * {@inheritdoc}
      */
     public function setCanClose(bool $canClose)
     {
         $this->canClose = $canClose;
-
-        return $this;
     }
 
     /**
-     * @return bool
+     * {@inheritdoc}
      */
     public function canOpen(): bool
     {
@@ -97,19 +93,7 @@ class Delimiter
     }
 
     /**
-     * @param bool $canOpen
-     *
-     * @return $this
-     */
-    public function setCanOpen(bool $canOpen)
-    {
-        $this->canOpen = $canOpen;
-
-        return $this;
-    }
-
-    /**
-     * @return bool
+     * {@inheritdoc}
      */
     public function isActive(): bool
     {
@@ -117,19 +101,15 @@ class Delimiter
     }
 
     /**
-     * @param bool $active
-     *
-     * @return $this
+     * {@inheritdoc}
      */
     public function setActive(bool $active)
     {
         $this->active = $active;
-
-        return $this;
     }
 
     /**
-     * @return string
+     * {@inheritdoc}
      */
     public function getChar(): string
     {
@@ -137,19 +117,7 @@ class Delimiter
     }
 
     /**
-     * @param string $char
-     *
-     * @return $this
-     */
-    public function setChar(string $char)
-    {
-        $this->char = $char;
-
-        return $this;
-    }
-
-    /**
-     * @return int|null
+     * {@inheritdoc}
      */
     public function getIndex(): ?int
     {
@@ -157,67 +125,47 @@ class Delimiter
     }
 
     /**
-     * @param int|null $index
-     *
-     * @return $this
+     * {@inheritdoc}
      */
-    public function setIndex(?int $index)
-    {
-        $this->index = $index;
-
-        return $this;
-    }
-
-    /**
-     * @return Delimiter|null
-     */
-    public function getNext(): ?self
+    public function getNext(): ?DelimiterInterface
     {
         return $this->next;
     }
 
     /**
-     * @param Delimiter|null $next
-     *
-     * @return $this
+     * {@inheritdoc}
      */
-    public function setNext(?self $next)
+    public function setNext(?DelimiterInterface $next)
     {
         $this->next = $next;
-
-        return $this;
     }
 
     /**
-     * @return int
+     * {@inheritdoc}
      */
-    public function getNumDelims(): int
+    public function getLength(): int
     {
-        return $this->numDelims;
+        return $this->length;
     }
 
     /**
-     * @param int $numDelims
-     *
-     * @return $this
+     * {@inheritdoc}
      */
-    public function setNumDelims(int $numDelims)
+    public function setLength(int $length)
     {
-        $this->numDelims = $numDelims;
-
-        return $this;
+        $this->length = $length;
     }
 
     /**
-     * @return int
+     * {@inheritdoc}
      */
-    public function getOrigDelims(): int
+    public function getOriginalLength(): int
     {
-        return $this->origDelims;
+        return $this->originalLength;
     }
 
     /**
-     * @return AbstractStringContainer
+     * {@inheritdoc}
      */
     public function getInlineNode(): AbstractStringContainer
     {
@@ -225,31 +173,17 @@ class Delimiter
     }
 
     /**
-     * @param AbstractStringContainer $node
-     *
-     * @return $this
+     * {@inheritdoc}
      */
-    public function setInlineNode(AbstractStringContainer $node)
-    {
-        $this->inlineNode = $node;
-
-        return $this;
-    }
-
-    /**
-     * @return Delimiter|null
-     */
-    public function getPrevious(): ?self
+    public function getPrevious(): ?DelimiterInterface
     {
         return $this->previous;
     }
 
     /**
-     * @param Delimiter|null $previous
-     *
-     * @return $this
+     * {@inheritdoc}
      */
-    public function setPrevious(?self $previous)
+    public function setPrevious(?DelimiterInterface $previous): DelimiterInterface
     {
         $this->previous = $previous;