+++ tests/unit/Delimiter/DelimiterTest.php
@@ -23,8 +23,8 @@ class DelimiterTest extends TestCase
 
         $delimiter = new Delimiter('*', 2, $node, true, false, null);
         $this->assertSame('*', $delimiter->getChar());
-        $this->assertSame(2, $delimiter->getNumDelims());
-        $this->assertSame(2, $delimiter->getOrigDelims());
+        $this->assertSame(2, $delimiter->getLength());
+        $this->assertSame(2, $delimiter->getOriginalLength());
         $this->assertSame($node, $delimiter->getInlineNode());
         $this->assertTrue($delimiter->canOpen());
         $this->assertFalse($delimiter->canClose());
@@ -32,8 +32,8 @@ class DelimiterTest extends TestCase
 
         $delimiter = new Delimiter('_', 1, $node, false, true, 7);
         $this->assertSame('_', $delimiter->getChar());
-        $this->assertSame(1, $delimiter->getNumDelims());
-        $this->assertSame(1, $delimiter->getOrigDelims());
+        $this->assertSame(1, $delimiter->getLength());
+        $this->assertSame(1, $delimiter->getOriginalLength());
         $this->assertSame($node, $delimiter->getInlineNode());
         $this->assertFalse($delimiter->canOpen());
         $this->assertTrue($delimiter->canClose());
@@ -49,15 +49,6 @@ class DelimiterTest extends TestCase
         $this->assertTrue($delimiter->canClose());
     }
 
-    public function testSetCanOpen()
-    {
-        $node = $this->createMock(AbstractStringContainer::class);
-        $delimiter = new Delimiter('*', 2, $node, true, false, null);
-
-        $delimiter->setCanOpen(false);
-        $this->assertFalse($delimiter->canOpen());
-    }
-
     public function testSetActive()
     {
         $node = $this->createMock(AbstractStringContainer::class);
@@ -68,27 +59,6 @@ class DelimiterTest extends TestCase
 
         $delimiter->setActive(false);
         $this->assertFalse($delimiter->isActive());
-    }
-
-    public function testSetChar()
-    {
-        $node = $this->createMock(AbstractStringContainer::class);
-        $delimiter = new Delimiter('*', 2, $node, true, false, null);
-
-        $delimiter->setChar('_');
-        $this->assertSame('_', $delimiter->getChar());
-    }
-
-    public function testSetIndex()
-    {
-        $node = $this->createMock(AbstractStringContainer::class);
-        $delimiter = new Delimiter('*', 2, $node, true, false, null);
-
-        $delimiter->setIndex(7);
-        $this->assertSame(7, $delimiter->getIndex());
-
-        $delimiter->setIndex(null);
-        $this->assertNull($delimiter->getIndex());
     }
 
     public function testSetNext()
@@ -103,25 +73,14 @@ class DelimiterTest extends TestCase
         $this->assertNull($delimiter->getNext());
     }
 
-    public function testSetNumDelims()
+    public function testSetLength()
     {
         $node = $this->createMock(AbstractStringContainer::class);
         $delimiter = new Delimiter('*', 2, $node, true, false, null);
 
-        $delimiter->setNumDelims(3);
-        $this->assertSame(3, $delimiter->getNumDelims());
-        $this->assertSame(2, $delimiter->getOrigDelims());
-    }
-
-    public function testSetInlineNode()
-    {
-        $node = $this->createMock(AbstractStringContainer::class);
-        $delimiter = new Delimiter('*', 2, $node, true, false, null);
-
-        $node2 = $this->createMock(AbstractStringContainer::class);
-
-        $delimiter->setInlineNode($node2);
-        $this->assertSame($node2, $delimiter->getInlineNode());
+        $delimiter->setLength(3);
+        $this->assertSame(3, $delimiter->getLength());
+        $this->assertSame(2, $delimiter->getOriginalLength());
     }
 
     public function testSetPrevious()