+++ CHANGELOG.md
@@ -4,6 +4,26 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+ - Made the `Delimiter` class final and extracted a new `DelimiterInterface`
+   - Modified most external usages to use this new interface
+ - Made the `DelimiterStack` class final
+ - Renamed three `Delimiter` methods:
+   - `getOrigDelims()` renamed to `getOriginalLength()`
+   - `getNumDelims()` renamed to `getLength()`
+   - `setNumDelims()` renamed to `setLength()`
+
+### Removed
+
+ - Removed unused `Delimiter` methods:
+   - `setCanOpen()`
+   - `setCanClose()`
+   - `setChar()`
+   - `setIndex()`
+   - `setInlineNode()`
+ - Removed fluent interface from `Delimiter` (setter methods now have no return values)
+
 ## [1.0.0-beta2] - 2019-05-27
 
 ### Changed